fix: move stack pattern client validator out of PatternChips
