+++ frontend/src/components/settings/PatternChips.tsx
@@ -2,15 +2,7 @@ import { forwardRef, useImperativeHandle, useState, type KeyboardEvent } from 'r
 import { Input } from '@/components/ui/input';
 import { Badge } from '@/components/ui/badge';
 import { X } from 'lucide-react';
-
-/** Client mirror of backend validateStackPatternForRedos. */
-export function validateStackPatternClient(pattern: string): string | null {
-  if (pattern.length > 200) return 'Pattern is too long (max 200 characters)';
-  const stars = (pattern.match(/\*/g) ?? []).length;
-  if (stars > 8) return 'Pattern has too many wildcards (max 8)';
-  if (/\*{4,}/.test(pattern)) return 'Pattern must not contain 4+ consecutive wildcards';
-  return null;
-}
+import { validateStackPatternClient } from './stackPatternClient';
 
 export interface PatternChipsHandle {
   /** Commit pending text if any; returns false when validation blocks save. */
